Split headers/sources
