+++ src/include/filters/AntiJitterFilter.h
